app/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==3.2.2
billiard==3.6.4.0
boto3==1.38.27
botocore==1.38.27
celery==5.2.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==6.5.0
cryptography==37.0.2
decorator==5.2.1
dependency-injector==4.48.3
dnspython==2.6.1
ecdsa==0.17.0
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==1.9.0
Faker==37.4.0
fastapi==0.115.12
fastapi-pagination==0.13.2
GeoAlchemy2==0.17.1
# greenlet==1.1.2
h11==0.16.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.5.3
loguru==0.6.0
Mako==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
more-itertools==10.7.0
moviepy
numpy==2.0.2
orjson==3.10.6
packaging==21.3
passlib==1.7.4
pillow==11.2.1
pluggy==1.0.0
proglog==0.1.12
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.18.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-cov==4.0.0
aiosqlite==0.19.0
httpx==0.27.0
pytest-forked==1.4.0
pytest-mock==3.9.0
pytest-xdist==2.5.0
python-dateutil==2.9.0.post0
python-dotenv
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.1
# PyYAML==6.0
redis==4.5.4
requests==2.28.1
rfc3986==1.5.0
rich==13.7.1
rsa==4.8
s3transfer==0.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.0.1
tqdm==4.67.1
typeguard==4.4.2
typer==0.12.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
wsproto==1.2.0
zipp==3.21.0
